Revert "Revert "Rename cli.common back to cli.utils""
This reverts commit 8fcf0578ed126eb9fadc2da3e784b82b041388cf.
