+++ script/publish_views
@@ -6,6 +6,3 @@ cd "$(dirname "$0")/.."
 
 ./bqetl bootstrap
 ./bqetl view publish "$@"
-
-wait
-trap - EXIT
