+++ script/validate_views
@@ -6,6 +6,3 @@ cd "$(dirname "$0")/.."
 
 ./bqetl bootstrap
 ./bqetl view validate "$@"
-
-wait
-trap - EXIT
