--- sql/moz-fx-data-test-project/test/dataset_metadata.yaml
@@ -1,10 +1,0 @@
-friendly_name: Test
-description: |-
-  Please provide a dataset description.
-dataset_base_acl: view
-user_facing: true
-labels: {}
-workgroup_access:
-- role: roles/bigquery.dataViewer
-  members:
-  - workgroup:mozilla-confidential
