--- sql/moz-fx-data-test-project/test/simple_view/view.sql
@@ -1,1 +1,0 @@
-SELECT 1
